\name{lab6pack-package}
\alias{lab6pack-package}
\alias{lab6pack}
\docType{package}
\title{
\packageTitle{lab6pack}
}
\description{
\packageDescription{lab6pack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab6pack}
\packageIndices{lab6pack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab6pack}

Maintainer: \packageMaintainer{lab6pack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
